[Refactor] generate curve generators
