+++ core/vm/privacy/bulletproof_test.go
@@ -7,54 +7,11 @@ import (
 	"math/big"
 	"os"
 	"testing"
-
-	"github.com/btcsuite/btcd/btcec"
 )
-
-func genECPrimeGroupKey(n int) CryptoParams {
-	// curValue := btcec.S256().Gx
-	// s256 := sha256.New()
-	gen1Vals := make([]ECPoint, n)
-	gen2Vals := make([]ECPoint, n)
-	// u := ECPoint{big.NewInt(0), big.NewInt(0)}
-	hx, _ := new(big.Int).SetString("50929b74c1a04954b78b4b6035e97a5e078a5a0f28ec96d547bfee9ace803ac0", 16)
-	hy, _ := new(big.Int).SetString("31d3c6863973926e049e637cb1b5f40a36dac28af1766968c30c2313f3a38904", 16)
-	ch := ECPoint{hx, hy}
-
-	gx, _ := new(big.Int).SetString("79be667ef9dcbbac55a06295ce870b07029bfcdb2dce28d959f2815b16f81798", 16)
-	gy, _ := new(big.Int).SetString("483ada7726a3c4655da4fbfc0e1108a8fd17b448a68554199c47d08ffb10d4b8", 16)
-	cg := ECPoint{gx, gy}
-
-	i := 0
-	for i < n {
-		gen2Vals[i] = ch.Mult(
-			big.NewInt(int64(i*2 + 1)),
-		)
-		gen1Vals[i] = cg.Mult(
-			big.NewInt(int64(i*2 + 2)),
-		)
-		i++
-	}
-
-	u := cg.Mult(
-		big.NewInt(int64(n + 3)),
-	)
-
-	return CryptoParams{
-		btcec.S256(),
-		btcec.S256(),
-		gen1Vals,
-		gen2Vals,
-		btcec.S256().N,
-		u,
-		n,
-		cg,
-		ch}
-}
 
 func TestInnerProductProveLen1(t *testing.T) {
 	fmt.Println("TestInnerProductProve1")
-	EC = NewECPrimeGroupKey(1)
+	EC = genECPrimeGroupKey(1)
 	a := make([]*big.Int, 1)
 	b := make([]*big.Int, 1)
 
@@ -104,7 +61,7 @@ func TestInnerProductProveLen2(t *testing.T) {
 
 func TestInnerProductProveLen4(t *testing.T) {
 	fmt.Println("TestInnerProductProve4")
-	EC = NewECPrimeGroupKey(4)
+	EC = genECPrimeGroupKey(4)
 	a := make([]*big.Int, 4)
 	b := make([]*big.Int, 4)
 
@@ -133,7 +90,7 @@ func TestInnerProductProveLen4(t *testing.T) {
 
 func TestInnerProductProveLen8(t *testing.T) {
 	fmt.Println("TestInnerProductProve8")
-	EC = NewECPrimeGroupKey(8)
+	EC = genECPrimeGroupKey(8)
 	a := make([]*big.Int, 8)
 	b := make([]*big.Int, 8)
 
@@ -170,7 +127,7 @@ func TestInnerProductProveLen8(t *testing.T) {
 
 func TestInnerProductProveLen64Rand(t *testing.T) {
 	fmt.Println("TestInnerProductProveLen64Rand")
-	EC = NewECPrimeGroupKey(64)
+	EC = genECPrimeGroupKey(64)
 	a := RandVector(64)
 	b := RandVector(64)
 
@@ -191,7 +148,7 @@ func TestInnerProductProveLen64Rand(t *testing.T) {
 
 func TestInnerProductVerifyFastLen1(t *testing.T) {
 	fmt.Println("TestInnerProductProve1")
-	EC = NewECPrimeGroupKey(1)
+	EC = genECPrimeGroupKey(1)
 	a := make([]*big.Int, 1)
 	b := make([]*big.Int, 1)
 
@@ -214,7 +171,7 @@ func TestInnerProductVerifyFastLen1(t *testing.T) {
 
 func TestInnerProductVerifyFastLen2(t *testing.T) {
 	fmt.Println("TestInnerProductProve2")
-	EC = NewECPrimeGroupKey(2)
+	EC = genECPrimeGroupKey(2)
 	a := make([]*big.Int, 2)
 	b := make([]*big.Int, 2)
 
@@ -239,7 +196,7 @@ func TestInnerProductVerifyFastLen2(t *testing.T) {
 
 func TestInnerProductVerifyFastLen4(t *testing.T) {
 	fmt.Println("TestInnerProductProve4")
-	EC = NewECPrimeGroupKey(4)
+	EC = genECPrimeGroupKey(4)
 	a := make([]*big.Int, 4)
 	b := make([]*big.Int, 4)
 
@@ -268,7 +225,7 @@ func TestInnerProductVerifyFastLen4(t *testing.T) {
 
 func TestInnerProductVerifyFastLen8(t *testing.T) {
 	fmt.Println("TestInnerProductProve8")
-	EC = NewECPrimeGroupKey(8)
+	EC = genECPrimeGroupKey(8)
 	a := make([]*big.Int, 8)
 	b := make([]*big.Int, 8)
 
@@ -305,7 +262,7 @@ func TestInnerProductVerifyFastLen8(t *testing.T) {
 
 func TestInnerProductVerifyFastLen64Rand(t *testing.T) {
 	fmt.Println("TestInnerProductProveLen64Rand")
-	EC = NewECPrimeGroupKey(64)
+	EC = genECPrimeGroupKey(64)
 	a := RandVector(64)
 	b := RandVector(64)
